requirements.txt
python-dotenv==1.0.0
MetaTrader5==5.0.45
google-generativeai==0.8.0
pandas==2.1.0
numpy<2.0.0
pandas-ta==0.3.14b0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0